reference/scenarios/aws-bedrock-agent/pyproject.toml
[project]
name = "aws-bedrock-agent-reference-test"
version = "0.0.0"
requires-python = ">=3.12"
dependencies = [
    "boto3==1.42.95",
    "genai-reference-shared",
]

[tool.uv.sources]
genai-reference-shared = { path = "../../shared", editable = true }

[tool.uv]
package = false
